BitbucketCloud: surround downloadJson() in try-catch
Move error handling wrapping of try-catch from call to downloadJson() in
convertPlainSubjectToHtml() to method downloadJson() itself, so that
other callers of method downloadJson() can also benefit from the error
handling.

Note that, with the way the `catch(e)` block is written, the returned
value from the function will be `undefined`.  A user will get a message
about the error in the log, but then the usage of `undefined` will throw
another error.