BitbucketCloud: style the link as a button
UI of Bitbucket Cloud on a commit page has several styled buttons:

    View source
    Approve
    •••
    Settings

The three dots shows a dropdown menu with "View raw commit" and "Stop
watching" actions.  Settings shows a modal dialog with settings for the
displaying of the diff and of the file view.  All these buttons have to
do with some action related to the commit or the commit's page.

Style the userscript's "Copy commit reference" link in the same way.
With new CSS classes, the workaround for aligning the icon with the text
(`margin-bottom: -7px;`) is no longer needed, so drop it.