GitHub: support transition PR page -> commit
The userscript matches GitHub pages with `/commit/` in the URL.
GitHub's fancy on-the-fly page reloading also supports seemless clicking
from a pull request's page to a commit.  Because the userscript isn't
loaded on a PR page, the "Copy commit reference" link won't be added to
the commit page that the user clicked to.

Match instead on the whole GitHub website.  This means that the
userscript will be loaded anywhere where an anchor pointing to a commit
page might be.  Even with GitHub's fancy page loading, the userscript's
`loadedSelector` will notice the moment when currently displayed GitHub
page changes into a commit page.  Then the recently introduced
optimization will kick in and only the GitHub's `getLoadedSelector()`
will be used, instead of concatenation of all supported hostings.

Bitbucket Cloud doesn't seem to be affected.  As far as I can tell,
a transition from a PR page to a commit page is always a full page
reload.