Cgit: show checkmark without date jump
The way copy link is laid out in the Cgit's header table, the dates
(author and commit) jump to the right during showing of the checkmark.

Add hacky CSS to avoid this jumping.  Using `position: absolute` is
incorrect, because we don't define left/top/right/bottom, but this
should be good enough for now.