Gitiles: fix tag pages
Tag pages on Gitiles have two sets of .Metadata and .MetadataMessage,
which are hard to distinguish with simple selectors.  Let's hope that
the HTML layout of Gitiles doesn't change much in the future, and just
use a `nth-child` hack to select the second .Metadata on tag pages,
which corresponds to the tagged commit, as opposed to the tag itself.
This might not work very well for lightweight tags, without a message.