GitLab: style checkmark like native tooltips
It is infeasible to mimic the native tooltips 100%.  So just add some
CSS to make the font, colors, paddings of userscript's checkmark tooltip
look almost like the native `.tooltip` class.  Can't use the CSS class
directly, because it will engage the native code for hiding the tooltips
and the element will disappear from the document.