Cgit: add support for newer versions of CGit
Site https://git.kernel.org/ runs not the most recent version of CGit.
In 2020 [1] CGit renamed CSS classes for SHA1 to new notation of oid,
short for object ID.  A newer build of CGit with this change included
runs on https://git.zx2c4.com/, where CGit is self-hosted.

Make CSS selectors used in class Cgit compatible with either versions of
CGit by simply listing the two posibilities separated by a comma.  A
possible alternative would be to use :is selector [2], but it seems
unnecessary -- just listing the copy-pasted rules is simple and readable
enough.

Add a `@match` entry for https://git.zx2c4.com/.

GitWeb lives in repository git.git, and there is also a CSS class .sha1
in GitWeb's code, but it is still called .sha1 as of Git 2.42.[3]

[1] commit 779631c (global: replace references to 'sha1' with 'oid',
    2020-10-20).
    https://git.zx2c4.com/cgit/commit/?id=779631c6dc23c15bbbf45a7c7ab9fffb619037b7
    https://github.com/zx2c4/cgit/commit/779631c6dc23c15bbbf45a7c7ab9fffb619037b7
    This commit is not included in any releases of CGit, as far as I can
    tell.  The latest release, CGit 1.2.3, was released in March 2020:
    Announcement: https://lists.zx2c4.com/pipermail/cgit/2020-March/004480.html
    Tag: https://git.zx2c4.com/cgit/tag/?h=v1.2.3
         https://github.com/zx2c4/cgit/releases/tag/v1.2.3
[2] https://stackoverflow.com/a/65960454/1083697
[3] https://github.com/git/git/blob/v2.42.0/gitweb/static/gitweb.css