GitHosting: use custom tag name for the button
Using anchor <a> with onclick handler for a user-clickable element
causes issues, because a click on an anchor usually causes a transition
in browser history.  This causes the userscript to misbehave on GitHub
and BitbucketCloud.

Make styling of the userscript's button easier by allowing subclasses to
override the actual HTML tag used by GitHosting for creating the button.
The default tag is <a href="#" ... />.  Add the "role" attribute for
accessibility.[1]

Choice of tag name per class:

  - GitHosting: <a href="#" .../> is the most common tag needed by
    subclasses, and it is also the least intrusive design-wise on most
    websites.  As explained above, it is not very compatible with
    single-page application (SPA) websites.
  - GitHub: <span> is the most neutral, not an <a>nchor that ruins
    popstate-related history navigation, not a <button> that has styling
    from browsers.
  - BitbucketCloud & GitLab: <button> for mimicry with native HTML and
    to avoid <a>nchor for BitbucketCloud.

[1] https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles