GitHosting: make all checkmarks `position: absolute`
Classes Cgit and GitLab override method createCheckmark() to make
'position' of the checkmark 'absolute'.  This is needed to accommodate
layout of the pages, which don't look good with the <span> tag of the
checkmark appearing right after the copy button.  This also facilitates
styling the checkmark in a way that is more native-like.

Style `checkmark` with `position: absolute` in class GitHosting to
replace custom styling done in classes Cgit and GitLab.  Remove extra
layer of a <span> containers for the checkmarks in Ggit and GitLab that
have `position: relative` and just apply `position: relative` to the big
`buttonContainer` in class GitHosting.

`position: absolute` messes up where the checkmark shows up for the
user.  Move styling `checkmark.style.left = 'calc(100% + 0.5rem)';` from
Cgit.createCheckmark to GitHosting.createCheckmark to make the styling
for all hostings OK-ish.