rename copy "link" to "button"
Using anchor <a> with onclick handler for a user-clickable element
causes issues, because a click on an anchor usually causes a transition
in browser history.

Prepare to replace <a> with a different tag by renaming all the symbols
(methods, functions, variables) from "link" to "button".  Some
subclasses of GitHosting, which use a look and feel of a button, already
refer to it as such.  Note that methods like #insertPrLinks in class
BitbucketServer are about proper hyperlinks.