GitHub: put button in the top right corner of commit view
On GitHub, the metadata (author, date, etc) is placed under the commit
message.  It means that the "Copy commit reference" button of the
userscript can change location based on the length of the commit
message.

Make the placement of the button more consistent by moving it to the top
of the commit message box.  In the new location of the button, the color
of the icon attracts too much attention.  Drop the removal of CSS class
'color-fg-muted' to make the icon less noticeable.

Because how GitHub's `@match` rule matches all pages of GitHub, the
userscript is also triggered on a commit view inside a pull request.
Changing the location of the button requires changing the selector in
method `getTargetSelector` in a way that is also applicable to such
pages in the PRs (the prior selector doesn't trigger).  The layout of
commit pages in pull requests is very close to the regular commit view,
so the "Copy reference button" makes sense there too.

With the above in mind, add support for commit pages in GitHub PRs.
Example URLs:

	https://github.com/rybak/atlassian-tweaks/pull/8/commits/d7cace2f9074d5311f066be33ea408b08a046946
	https://github.com/rybak/atlassian-tweaks/pull/11/commits/c73ec1c0788565dd4c7faac898a26a2198e5dcef
	https://github.com/junit-team/junit5/pull/3416/commits/3fad8c6c2a3829e2e329b334cd49b19f179d5f1f

Closes https://github.com/rybak/copy-commit-reference-userscript/issues/3
       "GitHub: design change idea"