.github & .gitlab: rename label "new-hosting-support" from "enhancement"
I've been creating more issues on GitHub, and label "enhancement" is too
vague.  I've split it into two: "feature" and "new-hosting-support",
since "new-hosting-support" is the most popular kind of issue so far.
At the moment issues are only present on GitHub and on Gogs:

  - https://github.com/rybak/copy-commit-reference-userscript/labels/feature
  - https://github.com/rybak/copy-commit-reference-userscript/labels/new-hosting-support
  - https://try.gogs.io/andrybak/copy-commit-reference-userscript/issues?type=all&state=open&labels=7433&milestone=0&assignee=0

Migrate the issue templates for GitHub and GitLab (no analogue for Gogs
or other mirrors yet) to the new label "new-hosting-support".

By the way, GitLab's Markdown has special syntax for linking to labels:

 - ~new-hosting-support
 - ~feature

GitHub, on the other hand, doesn't seem to have a special Markdown
syntax for labels, but it does convert full URLs to the labels (see
above) in Markdown to fancy clickable oval-shaped buttons, similar to
the ones produced by GitLab.