docs: add Greasy Fork description file
Automate maintaining descriptions for the userscripts on Greasy Fork by
storing a hosting-agnostic description in Markdown format in the
repository.

I don't have a lot of experience using this feature of Greasy Fork.  So
far, it _seems_ that Greasy Fork only updates the descriptions if there
were updates in the source code.  Because of this seeming restriction,
this approach might turn out to not be very useful.