docs: move CHANGELOG.md to OLD_CHANGELOG.md
Similarly to commit 3aed20a (drop copy-commit-reference.user.js,
2023-10-06), file CHANGELOG.md is no longer actual.  However, it is a
useful reference, so move it to the directory `Documentation/` under
name `OLD_CHANGELOG.md`.  The new name should make it obvious that this
file isn't actual anymore and it should stop GitLab from generating a
special button/badge for it at the top of the repository page.