GitLab: add gitlab.gnome.org as a `@match` entry
KDE's invent.kde.org is in `@match` rules of the userscript for GitLab,
because it was there in the big userscript before the split into
standalone per-hosting scripts.  KDE's GitLab instance was in the
`@match` in the big userscript simply because I'm currently using KDE
and I sometimes use their instance of GitLab to look up code.  So it
made sense to use it just to have some comparison to gitlab.com -- they
might have different versions, different look and feel theme, etc.

Today I found out that Gnome is also using a GitLab instance.  Add their
instance as a `@match` rule to gitlab-copy-commit-reference.user.js.