GitLab: support issue numbers in subject
Override method convertPlainSubjectToHtml of class GitHostings to add
support of issue numbers in subject line of commits on GitLab.

This wasn't implemented before because I myself haven't used issues on
gitlab.com and KDE project doesn't use issues on invent.kde.org, because
they use Bugzilla [1] for issue tracking.  But after adding a `@match`
entry for gitlab.gnome.org in the previous commit, I quickly stumbled
upon a commit with issue number in the subject [2].

GitLab uses hash sign for issue references and exclamation marks for
merge requests,[3] unlike GitHub and Gitea, where issue numbers and pull
request numbers share the same space and are both referenced with '#'.
By default GitLab puts the reference to a Merge Request into the body of
a merge commit message.[4] I haven't been able to find a commit with
exclamation mark reference to an MR in the subject line of the commit
message.  So support only issue numbers for now.  If MRs are referenced
in the subject lines in practice, we would implement support for
exclamation mark references as well.

[1] https://bugs.kde.org/
[2] https://gitlab.gnome.org/ilhooq/bookup/-/commit/7bef5b7ce54ac84451eff99f65cd58fc283dffe5
[3] https://docs.gitlab.com/ee/user/markdown.html#gitlab-specific-references
[4] https://docs.gitlab.com/ee/user/project/merge_requests/commit_templates.html#default-template-for-merge-commits
    Quote:

    > The default template for merge commit messages is:
    >
    >     Merge branch '%{source_branch}' into '%{target_branch}'
    >
    >     %{title}
    >
    >     %{issues}
    >
    >     See merge request %{reference}