GitWeb: use document.createElement for barSeparator
Userscript gitweb-copy-commit-reference.user.js adds a bar separator
before the copy commit reference button for consistency with the
GitWeb's native UI.  It is done using function htmlToElement,[1] which
is way too heavy and complicated for creating a single tag with simple
contents and a single CSS class.

Stop using function htmlToElement.  Migrate to version of the library
copy-commit-reference-lib.js, which no longer has the function
htmlToElement.

[1] https://stackoverflow.com/a/35385518/1083697