README: add Codeberg mirror
The Forgejo instance which is listed as a mirror for this repository,
i.e. next.forgejo.org, is a demo instance.  I have finally managed to
create an account on codeberg.org (they don't like Gmail).  Codeberg
itself runs on Forgejo.

Add Codeberg as a source code mirror to the lists in documentation.
Replace Forgejo in the lists of issue trackers with a corresponding
Codeberg link.