Gogs: better compatibility with older versions
Older versions of Gogs don't have the CSS class ".commit-message".
Update function getCommitMessage() to take the commit message out of the
header <h3> in the container of the commit message, which has the unique
CSS class ".top".

Update functions getIssuesUrl and convertPlainSubjectToHtml to handle
cases where the issues are turned off in the repository.  Commits with
issue numbers in the message are possible, if the repository is a simple
mirror or a fork, which doesn't use the issues feature.