GitHub: update HTML layout
A recent update to GitHub's HTML layout broke the userscript.

Update methods getTargetSelector, wrapButtonContainer,
addButtonContainerToTarget, and wrapButton to accommodate these changes.

For pull request pages, the button border looks a bit weird due to
closeness to the "< Prev | Next >" buttons.  This is good enough for
now, but it would be nice to untangle that.