GitHub, GitLab, Gitea/Forgejo, Gogs: escape HTML in convertPlainSubjectToHtml
A bug in version 0.4-alpha of copy-commit-reference-lib.js caused
incorrect HTML to be generated when issues or pull requests are linked
in the commit message.  This affected classes GitHub, GitLab, Gitea, and
Gogs.

Migrate the corresponding scripts to the new version of the library with
the fix.  Update HTML handling in the method convertPlainSubjectToHtml
of the classes according to the new documentation of the class
GitHosting in the library code.