Gitea/Forgejo, GitLab, Gogs: mark convertPlainSubjectToHtml as async
In commit [1], methods `convertPlainSubjectToHtml` in classes Gitea,
GitLab, and Gogs gained a `await` call to the method
`super.convertPlainSubjectToHtml`.  Because these methods are not marked
as `async` themselves, it causes an error, like this:

	SyntaxError: await is only valid in async functions and the top
	level bodies of modules at undefined:161:23.

Add the missing `async` qualifiers to these methods.  The corresponding
method in class GitHub already had the `async` modifier, so it wasn't
affected.

[1] d4905c2 (GitHub, GitLab, Gitea/Forgejo, Gogs: escape HTML in
    convertPlainSubjectToHtml, 2024-12-03)