GitLab: addButtonContainerToTarget: update selector
CSS classes in HTML layout on GitLab changed, so the variable
`authoredSpanTag` became `null` in method `addButtonContainerToTarget`
of class GitLab.

Switch to a selector without CSS class names, just taking the first
<span> tag in the target HTML element.  At the time of writing the
relevant HTML looks like this:

	<div class="header-main-content"><!-- target -->
		<strong>Commit ...</strong>
		<button>...</button><!-- native copy commit hash button -->
		<!-- `buttonContainer` is inserted here -->
		<span class="gl-hidden @sm/panel:gl-inline">authored</span>
		<time class="js-timeago" ...>...</time>
		<span>by</span>
		<strong>...</strong>
		<span>...</span>
	</div>

We could use the `gl-hidden` CSS class, but it might break in the
future.